backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
av==16.0.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.2
ctranslate2==4.6.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
faster-whisper==1.2.0
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.9.0
google-auth==2.42.1
google-genai==1.48.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
numpy==2.3.4
onnxruntime==1.23.2
packaging==25.0
passlib==1.7.4
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyreadline3==3.5.4
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
